Adding file that can iteratively process a directory
Added upright_stabalized_iterative.py.  This file is an optimized version of upright_stabalized.py.  It will pull all file directories to process as well as output files from the .env file.  Also logging was added, the file location and name come from a value in the .env file.  Updated the .gitignore to ignore outputs directories and .venv.